pyproject.toml
[tool.poetry]
name = "anna"
version = "0.1.0"
description = ""
authors = ["mika <mika.griebsch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.4"
Django = "^4.0.4"
django-bootstrap5 = "^22.1"
django-jazzmin = "^2.5.0"
django-quill-editor = "^0.1.40"
django-storages = "^1.12.3"
django-user-agents = "^0.4.0"
Pillow = "^9.1.1"
python-dateutil = "^2.8.2"
python-decouple = "^3.6"
boto3 = "^1.23.4"
html2text = "^2020.1.16"
gunicorn = "^20.1.0"
django-secrets = "^1.2.2"
django-extensions = "^3.1.5"
ipdb = "^0.13.9"
colorama = "^0.4.4"
fabric3 = {git = "https://github.com/kakulukia/fabric3.git"}
icecream = "^2.1.2"
django-undeletable = "^1.1.2"
black = "^22.3.0"
isort = "^5.10.1"
mutagen = "^1.45.1"
opencv-python = "^4.6.0.66"
django-debug-toolbar = "^3.4.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 95
exclude = '''
/(
  \.direnv
  | \.venv
  | migrations
)/
'''

[tool.isort]
profile = "black"
